Create the libAPU for the host
We want to use the APU to offload the CPU for NN/ML operations.
In order to make easier to use and implement in different NN/ML
framework, create a library.
This mostly provides RPC functions. A function implemented and
registered on the APU could be executed from the host.

Note:
Currently, this lacks support of build system.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>