Update the resource table and the linker script
The resource table contains an unused entry so we could remove it.
In addition, update the linker script to "reserve" the memory buffer
also defined in the resource table. The goal is to make easier to
update the linker script and the resource table.

Note:
There are still a lot of work to do to make the resource table
and the linker script more scalable.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>